+++ packages/terra-framework-docs/src/terra-dev-site/doc/navigation-side-menu/example/DrillInExample.jsx
@@ -39,7 +39,12 @@ class DrillInDefault extends React.Component {
             key: 'submenu3', text: 'Hospital Accommodations', isDisabled: true,
           },
           { key: 'submenu4', text: 'Hospital Careers' },
-          { key: 'submenu5', text: 'Hospital Info', childKeys: [] },
+          {
+            key: 'submenu5',
+            text: 'Hospital Info',
+            childKeys: [],
+            customStatusDisplay: <span>No results for Hospital Info</span>,
+          },
           {
             key: 'subsubmenu1', text: 'Imaging', id: 'test-item-2',
           },