+++ packages/terra-navigation-side-menu/tests/wdio/navigation-side-menu-spec.js
@@ -119,10 +119,29 @@ Terra.describeViewports('DrillIn View', ['medium'], () => {
       browser.keys(['ArrowDown']);
       Terra.validates.element('second drill-in item focused', { selector: '#root' });
     });
+  });
 
-    it('Navigate into folder with no items', () => {
-      browser.keys(['ArrowDown', 'ArrowDown', 'ArrowDown', 'ArrowRight']);
-      Terra.validates.element('displays no results status', { selector: '#root' });
+  describe('DrillIn status display for folder with no items', () => {
+    it('Navigates into folder with no items', () => {
+      browser.url('/#/raw/tests/cerner-terra-framework-docs/navigation-side-menu/drill-in');
+      browser.keys(['Tab', 'ArrowDown', 'ArrowRight', 'ArrowDown', 'ArrowDown', 'ArrowDown', 'ArrowRight',
+        'ArrowDown', 'ArrowDown', 'ArrowDown', 'ArrowRight']);
+      Terra.validates.element('displays no results status-view', { selector: '#root' });
+    });
+
+    it('Should display custom status for folder with no items', () => {
+      browser.url('/#/raw/tests/cerner-terra-framework-docs/navigation-side-menu/drill-in');
+      browser.keys(['Tab', 'ArrowDown', 'ArrowDown', 'ArrowDown', 'ArrowDown', 'ArrowDown', 'ArrowRight']);
+      Terra.validates.element('displays no results custom status', { selector: '#root' });
+    });
+
+    it('Should display custom status when Loading is true and data when loading is false', () => {
+      browser.url('/#/raw/tests/cerner-terra-framework-docs/navigation-side-menu/drill-in');
+      browser.keys(['Tab', 'ArrowDown', 'ArrowRight', 'ArrowDown', 'ArrowDown', 'ArrowDown']);
+      browser.keys(['ArrowRight', 'ArrowDown', 'ArrowDown', 'ArrowRight']);
+      Terra.validates.element('displays loading status', { selector: '#root' });
+      browser.pause(1000);
+      Terra.validates.element('displays data after loading', { selector: '#root' });
     });
   });
 });