+++ packages/terra-framework-docs/CHANGELOG.md
@@ -5,7 +5,7 @@
 * Changed
   * Enabled `close` button in examples.
 
-* Changed
+* Updated
   * Updated example for drill-in view under `terra-navigation-side-menu`.
 
 ## 1.89.1 - (May 20, 2024)